requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
psycopg2-binary
sentence-transformers
faiss-cpu
numpy
requests
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
